+++ src/parser/mod.rs
@@ -1,4 +1,4 @@
-pub use self::main::{parse_object, parse_whois_response};
+pub use api::{parse_object, parse_whois_response};
 
-mod component;
-mod main;
+mod api;
+mod core;
